+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Theme handling via `cookieStgore`
+
 ## [v1.0.7] - 2020-11-28
 
 ### Changed